apps/services/servers/dashboardMs/requirements.txt
apache-beam==2.48.0
apache-flink==1.18.0
apache-flink-libraries==1.18.0
avro-python3==1.9.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
confluent-kafka==2.3.0
crcmod==1.7
dill==0.3.1.1
dnspython==2.4.2
docopt==0.6.2
fastavro==1.9.1
fasteners==0.19
find-libpython==0.3.1
Flask==3.0.0
Flask-Cors==4.0.0
grpcio==1.60.0
gunicorn==21.2.0
hdfs==2.7.3
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.4
objsize==0.6.1
orjson==3.9.10
packaging==23.2
pandas==2.1.4
pemja==0.3.0
proto-plus==1.23.0
protobuf==4.23.4
py4j==0.10.9.7
pyarrow==11.0.0
pydot==1.4.2
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
zstandard==0.22.0
